4.21

+++ controllers/blogs.js
@@ -47,6 +47,10 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== userId.toString()) {
+    return response.status(401).json( { error: 'Unauthorized' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
